ver2

+++ src/main/java/com/dsr/navigationapp/controllers/DirectoryController.java
@@ -3,11 +3,13 @@ package com.dsr.navigationapp.controllers;
 import com.dsr.navigationapp.filesystem.DirectoryInfo;
 import com.dsr.navigationapp.filesystem.DirectoryNavigation;
 import com.dsr.navigationapp.models.FileInfo;
+import com.dsr.navigationapp.models.requests.rest.MoveCopyRequest;
 import com.dsr.navigationapp.models.requests.rest.RenameRequestBody;
 import com.dsr.navigationapp.services.NavigationService;
 import com.dsr.navigationapp.models.responses.rest.BooleanResponseBody;
 import com.dsr.navigationapp.models.responses.rest.DeleteResponseBody;
 import com.dsr.navigationapp.models.responses.rest.PathContentResponse;
+import com.dsr.navigationapp.utils.ValidatorClass;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -17,56 +19,78 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/api/v1/web/")
 public class DirectoryController {
     @GetMapping(value = "/directories")
-    public ResponseEntity<PathContentResponse> get(
+    public ResponseEntity<?> get(
             @RequestParam(value = "path") String path,
             @RequestParam(value = "page", defaultValue = "1", required = false) Integer page,
             @RequestParam(value = "size", required = false) Integer size
     ) {
+        var isValid = ValidatorClass.validatePath(path);
+        if (!isValid) {
+            return ResponseEntity.badRequest().body("Not valid path");
+        }
         path = DirectoryNavigation.getInstance().getFullPath(path);
         return ResponseEntity.ok(NavigationService.getPathContent(path, page, size));
     }
 
     @GetMapping(value = "/directories/info")
-    public ResponseEntity<FileInfo> getInfo(
+    public ResponseEntity<?> getInfo(
             @RequestParam(value = "path") String path
     ) {
+        var isValid = ValidatorClass.validatePath(path);
+        if (!isValid) {
+            return ResponseEntity.badRequest().body("Not valid path");
+        }
         path = DirectoryNavigation.getInstance().getFullPath(path);
         return ResponseEntity.ok(DirectoryInfo.getFileInfo(path));
     }
 
     @DeleteMapping(value = "/directories")
-    public ResponseEntity<DeleteResponseBody> delete(
+    public ResponseEntity<?> delete(
             @RequestParam(value = "path") String path
     ) {
+        var isValid = ValidatorClass.validatePath(path);
+        if (!isValid) {
+            return ResponseEntity.badRequest().body("Not valid path");
+        }
         path = DirectoryNavigation.getInstance().getFullPath(path);
         return ResponseEntity.ok(NavigationService.removeFile(path));
     }
 
     @PutMapping(value = "/directories")
-    public ResponseEntity<BooleanResponseBody> rename(
+    public ResponseEntity<?> rename(
             @RequestBody RenameRequestBody requestBody
     ) {
+        var isValid = ValidatorClass.validatePath(requestBody.getPath());
+        if (!isValid) {
+            return ResponseEntity.badRequest().body("Not valid path");
+        }
         var path = DirectoryNavigation.getInstance().getFullPath(requestBody.getPath());
         return ResponseEntity.ok(NavigationService.rename(path, requestBody.getNewName()));
     }
 
     @PutMapping(value = "/directories/copy")
-    public ResponseEntity<BooleanResponseBody> copy(
-            @RequestParam(value = "path") String path,
-            @RequestParam(value = "destination") String destination
+    public ResponseEntity<?> copy(
+            @RequestBody MoveCopyRequest body
     ) {
-        path = DirectoryNavigation.getInstance().getFullPath(path);
-        destination = DirectoryNavigation.getInstance().getFullPath(destination);
+        var isValid = ValidatorClass.validatePath(body.getPath());
+        if (!isValid) {
+            return ResponseEntity.badRequest().body("Not valid path");
+        }
+        var path = DirectoryNavigation.getInstance().getFullPath(body.getPath());
+        var destination = DirectoryNavigation.getInstance().getFullPath(body.getDestination());
         return ResponseEntity.ok(NavigationService.copy(path, destination));
     }
 
     @PutMapping(value = "/directories/move")
-    public ResponseEntity<BooleanResponseBody> move(
-            @RequestParam(value = "path") String path,
-            @RequestParam(value = "destination") String destination
+    public ResponseEntity<?> move(
+            @RequestBody MoveCopyRequest body
     ) {
-        path = DirectoryNavigation.getInstance().getFullPath(path);
-        destination = DirectoryNavigation.getInstance().getFullPath(destination);
+        var isValid = ValidatorClass.validatePath(body.getPath());
+        if (!isValid) {
+            return ResponseEntity.badRequest().body("Not valid path");
+        }
+        var path = DirectoryNavigation.getInstance().getFullPath(body.getPath());
+        var destination = DirectoryNavigation.getInstance().getFullPath(body.getDestination());
         return ResponseEntity.ok(NavigationService.move(path, destination));
     }
 }
